change id's type

+++ src/main/java/com/fedag/CSR/repository/UserRepository.java
@@ -5,6 +5,8 @@ import com.fedag.CSR.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, BigDecimal> {
 }
